backend/requirements.txt
pipecat-ai[google,silero,websocket]>=0.0.105
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
pydantic-settings>=2.0.0
loguru>=0.7.0
chromadb>=1.0.0
httpx>=0.27.0
